Move package to org.http4s.scalaxml1 (#1)
